requirements.txt
beautifulsoup4>=4.4.0
python-dateutil>=2.4.2
requests>=2.7.0
